+++ loops/loop.c
@@ -20,7 +20,7 @@ void endLoop() {
     }
 
     char *body = loop_mode->body;
-    int iter = loop_mode->iter;
+    long long int iter = loop_mode->iter;
     enum LoopType loop_type = loop_mode->type;
     char *array_name = loop_mode->array;
     char *element_name = loop_mode->element.name;
@@ -36,13 +36,13 @@ void endLoop() {
     switch (loop_type)
     {
         case TIMESDO:
-            for (int i = 0; i < iter; i++) {
+            for (long long int i = 0; i < iter; i++) {
                 injectCode(body, INIT_PROGRAM);
             }
             break;
         case FOREACH:
             array = getSymbol(array_name);
-            for (int i = 0; i < array->children_count; i++) {
+            for (long long int i = 0; i < array->children_count; i++) {
                 Symbol* child = array->children[i];
                 child->name = element_name;
                 injectCode(body, INIT_PROGRAM);
@@ -51,7 +51,7 @@ void endLoop() {
             break;
         case FOREACH_DICT:
             array = getSymbol(array_name);
-            for (int i = 0; i < array->children_count; i++) {
+            for (long long int i = 0; i < array->children_count; i++) {
                 Symbol* child = array->children[i];
                 char *key = malloc(1 + strlen(child->key));
                 char *element_key_copy = malloc(1 + strlen(element_key));
