+++ loops/loop.h
@@ -15,8 +15,8 @@ typedef struct {
 
 typedef struct {
     enum LoopType type;
-    int iter;
-    int nested_counter;
+    unsigned short int iter;
+    unsigned short int nested_counter;
     char *array;
     LoopElement element;
     char *body;
